% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skim_data.r
\name{skim_data}
\alias{skim_data}
\title{Skim Data}
\usage{
skim_data(x)
}
\arguments{
\item{x}{A data frame to be skimmed.}
}
\value{
A data frame with the skimmed data. Each row corresponds to a variable in the input data.frame and contains the following columns:
\itemize{
\item variable: the name of the variable
\item type: the type of the variable
\item range: the range of the variable if it is numeric, the number of levels if it is a factor, or the number of unique values if it is a character
\item missing: the number of missing values in the variable
\item distribution: the distribution of the variable if it is numeric and the \pkg{skimr} package is available
}
}
\description{
This function skims a data frame. It provides information about each variable such as its type, range, missing values, and distribution (if the \pkg{skimr} package is available).
}
\examples{
\dontrun{
skim_data(df)
}
}
